Removed old descriptors, moved journal module to top level
