+++ libs/ec-slimloader-state/src/flash/mock.rs
