--- libs/ec-slimloader-state/src/journal/mod.rs
@@ -1,4 +1,0 @@
-//! Journal for the EC Slimloader containing [state::State].
-
-pub mod flash;
-pub mod state;
