+++ libs/ec-slimloader-state/src/state.rs
